+++ src/lib/logging.h
@@ -24,25 +24,6 @@
 #include <stdlib.h>
 #include "util.h"
 
-enum log_state {
-	LS_INIT, // Initial state
-	LS_CONF, // Running configuration scripts
-	LS_PLAN, // Generating plan
-	LS_DOWN, // Downloading needed packages
-	LS_PREUPD, // Running updater's preupdate hooks
-	// These are states in critical section
-	LS_UNPACK, // Unpacking downloaded packages
-	LS_CHECK, // Checking for files collisions and more
-	LS_INST, // Running preinst scripts and merging files to filesystem
-	LS_POST, // Running postinst scripts
-	LS_REM, // Removing leftover files
-	LS_CLEANUP, // Cleaning up control files
-	// End of critical section
-	LS_POSTUPD, // Running updater's postupdate hooks
-	LS_EXIT, // Updater exit
-	LS_FAIL, // Detected failure, exiting.
-};
-
 enum log_level {
 	LL_DISABLE,
 	LL_DIE,
@@ -87,13 +68,6 @@ struct log_buffer {
 };
 // Initialize log buffer (if would_log(level)
 void log_buffer_init(struct log_buffer *buf, enum log_level level) __attribute__((nonnull));
-
-// Sets if state and error should be dumped into files in /tmp/updater-state directory
-void set_state_log(bool state_log);
-// Log state change of updater
-void update_state(enum log_state);
-// In the full updater mode, dump the error into /tmp/update-state/error
-void err_dump(const char *msg) __attribute__((nonnull));
 
 void log_syslog_level(enum log_level level);
 void log_syslog_name(const char *name);