+++ src/lib/lua/transaction.lua
@@ -42,16 +42,7 @@ local changelog = require "changelog"
 local DBG = DBG
 local WARN = WARN
 local INFO = INFO
-local LS_UNPACK = LS_UNPACK
-local LS_INST = LS_INST
-local LS_CHECK = LS_CHECK
-local LS_INST = LS_INST
-local LS_POST = LS_POST
-local LS_REM = LS_REM
-local LS_CLEANUP = LS_CLEANUP
-local update_state = update_state
 local sync = sync
-local log_event = log_event
 
 module "transaction"
 
@@ -69,7 +60,6 @@ end
 
 -- Stages of the transaction. Each one is written into the journal, with its results.
 local function pkg_unpack(operations, status)
-	update_state(LS_UNPACK)
 	INFO("Unpacking download packages")
 	local dir_cleanups = {}
 	--[[
@@ -141,7 +131,6 @@ local function pkg_unpack(operations, status)
 end
 
 local function pkg_collision_check(status, to_remove, to_install)
-	update_state(LS_CHECK)
 	INFO("Checking for file collisions between packages")
 	local collisions, early_remove, removes = backend.collision_check(status, to_remove, to_install)
 	if next(collisions) then
@@ -176,7 +165,6 @@ local function changelog_end(curchangelog)
 end
 
 local function pkg_move(status, plan, early_remove, errors_collected, curchangelog)
-	update_state(LS_INST)
 	INFO("Running pre-install and pre-rm scripts and merging packages to root file system")
 	-- Prepare table of not installed confs for config stealing
 	local installed_confs = backend.installed_confs(status)
@@ -194,7 +182,6 @@ local function pkg_move(status, plan, early_remove, errors_collected, curchangel
 	local upgraded_packages = {}
 	for _, op in ipairs(plan) do
 		if op.op == "install" then
-			log_event("I", op.control.Package .. " " .. op.control.Version)
 			-- Unfortunately, we need to merge the control files first, otherwise the maintainer scripts won't run. They expect to live in the info dir when they are run. And we need to run the preinst script before merging the files.
 			backend.pkg_merge_control(op.dir .. "/control", op.control.Package, op.control.files)
 			if utils.multi_index(status, op.control.Package, "Status", 3) == "installed" then
@@ -214,7 +201,6 @@ local function pkg_move(status, plan, early_remove, errors_collected, curchangel
 			status[op.control.Package] = op.control
 			backend.pkg_update_alternatives(status, op.control.Package)
 		elseif op.op == "remove" and utils.arr2set(utils.multi_index(status, op.name, 'Status') or {})['installed'] then
-			log_event("R", op.name)
 			utils.table_merge(all_configs, status[op.name].Conffiles or {})
 			local cfiles = status[op.name].Conffiles or {}
 			for f in pairs(cfiles) do
@@ -246,10 +232,8 @@ end
 local function pkg_scripts(status, plan, removes, to_install, errors_collected, all_configs, upgraded_packages, curchangelog)
 	-- Clean up the files from removed or upgraded packages
 	INFO("Removing packages and leftover files")
-	update_state(LS_REM)
 	backend.pkg_cleanup_files(removes, all_configs)
 	-- Run post install and remove scripts
-	update_state(LS_POST)
 	INFO("Running post-install and post-rm scripts")
 	for _, op in ipairs(plan) do
 		if op.op == "install" then
@@ -262,7 +246,6 @@ local function pkg_scripts(status, plan, removes, to_install, errors_collected, 
 end
 
 local function pkg_cleanup(status)
-	update_state(LS_CLEANUP)
 	INFO("Cleaning up control files")
 	backend.control_cleanup(status)
 	backend.status_dump(status)